feat(console-ui): move leaderboard to standalone /leaderboard tab
Extract the leaderboard out of the monolithic stats page into a modular
src/components/leaderboard/ package (types, useLeaderboard hook, format
helpers + tests, TotalsStrip, PodiumCard, RankingsTable, Controls, thin
LeaderboardContent orchestrator), add a /leaderboard route and sidebar
nav item, and make the Stats page overview-only. Behavior-preserving
port; same /api/leaderboard + /api/network/totals proxies.
