+++ console-ui/src/app/leaderboard/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import { LeaderboardContent } from "@/components/leaderboard/LeaderboardContent";
+
+export default function LeaderboardPage() {
+  return <LeaderboardContent />;
+}
