+++ console-ui/src/components/Sidebar.tsx
@@ -16,6 +16,7 @@ import {
   Code,
   Activity,
   Coins,
+  Trophy,
   LogOut,
   Sun,
   Moon,
@@ -70,6 +71,7 @@ export function Sidebar() {
         {[
           { href: "/", icon: MessageSquare, label: "Chat" },
           { href: "/stats", icon: Activity, label: "Stats" },
+          { href: "/leaderboard", icon: Trophy, label: "Leaderboard" },
           { href: "/providers", icon: Server, label: "Provider Dashboard" },
           { href: "/earn", icon: Coins, label: "Earn" },
           { href: "/api-console", icon: Code, label: "API" },
